get rid of circular dependancies
