+++ src-tauri/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "sprocket_pan"
-version = "0.0.0"
-description = "A Tauri App"
+version = "1.0.0"
+description = "Sprocket Pan"
 authors = ["Itai Fish"]
 license = ""
 repository = ""
@@ -13,7 +13,14 @@ edition = "2021"
 tauri-build = { version = "1.4", features = [] }
 
 [dependencies]
-tauri = { version = "1.4", features = [ "dialog-all", "os-all", "path-all", "fs-all", "shell-open"] }
+tauri = { version = "1.4", features = [
+	"dialog-all",
+	"os-all",
+	"path-all",
+	"fs-all",
+	"shell-open",
+	"devtools",
+] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 tauri-plugin-window-state = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v1" }
