+++ src-tauri/tauri.conf.json
@@ -8,7 +8,7 @@
 	},
 	"package": {
 		"productName": "SprocketPan",
-		"version": "0.0.0"
+		"version": "1.0.0"
 	},
 	"tauri": {
 		"allowlist": {
