+++ src/components/molecules/file-system/RequestFileSystem.tsx
@@ -12,7 +12,6 @@ import {
 import { EndpointRequest } from '../../../types/application-data/application-data';
 import { useContext, useState } from 'react';
 import TextSnippetIcon from '@mui/icons-material/TextSnippet';
-import { ApplicationDataContext, TabsContext } from '../../../App';
 import { keepStringLengthReasonable } from '../../../utils/string';
 import { tabsManager } from '../../../managers/TabsManager';
 import { MoreVert } from '@mui/icons-material';
@@ -20,6 +19,7 @@ import { applicationDataManager } from '../../../managers/ApplicationDataManager
 import FolderCopyIcon from '@mui/icons-material/FolderCopy';
 import DeleteForeverIcon from '@mui/icons-material/DeleteForever';
 import { AreYouSureModal } from '../../atoms/modals/AreYouSureModal';
+import { ApplicationDataContext, TabsContext } from '../../../managers/GlobalContextManager';
 
 export function RequestFileSystem({ request }: { request: EndpointRequest }) {
 	const tabsContext = useContext(TabsContext);
