+++ src/components/molecules/tabs/tab-props.ts
@@ -1,0 +1,3 @@
+export interface TabProps {
+	id: string;
+}
